fix(offline): encoding map value with codegen
ensure each row encoded with distinct jit instance, work-around to #3748
